recommender-ai-service/requirements.txt
django
djangorestframework
scikit-surprise
requests
pandas
numpy<2
